Detail Screens Coded

+++ app/src/main/java/com/example/appcentcase/Model/ConsolidatedWeather.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class ConsolidatedWeather(
     @SerializedName("air_pressure")
-    val airPressure: Int,
+    val airPressure: Double,
     @SerializedName("applicable_date")
     val applicableDate: String,
     @SerializedName("created")
